move field to another file
